Silence some GCC warnings

+++ common/lib_tree_model_adapter.cpp
@@ -416,7 +416,8 @@ wxDataViewColumn* LIB_TREE_MODEL_ADAPTER::doAddColumn( const wxString& aHeader, 
 
     wxDataViewColumn* col = new wxDataViewColumn(
             translatedHeader, new LIB_TREE_RENDERER(), index, m_colWidths[aHeader], wxALIGN_NOT,
-            wxDATAVIEW_CELL_INERT | wxDATAVIEW_COL_RESIZABLE | wxDATAVIEW_COL_SORTABLE );
+            wxDATAVIEW_CELL_INERT | static_cast<int>( wxDATAVIEW_COL_RESIZABLE )
+                    | static_cast<int>( wxDATAVIEW_COL_SORTABLE ) );
     m_widget->AppendColumn( col );
 
     col->SetMinWidth( headerMinWidth.x );
